moved all helper functions used in getRoutes to HelperFunctions file
